Initial setup for push support in test app
