+++ DemoApp/StreamChat/StreamChatWrapper+DemoApp.swift
@@ -1,0 +1,24 @@
+//
+// Copyright © 2022 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+import StreamChat
+import StreamChatUI
+
+extension StreamChatWrapper {
+    
+    func configureUI() {
+        // Customize UI configuration
+        Components.default.messageListDateSeparatorEnabled = true
+        Components.default.messageListDateOverlayEnabled = true
+
+        // Customize UI components
+        Components.default.channelListRouter = DemoChatChannelListRouter.self
+        Components.default.channelVC = DemoChatChannelVC.self
+        Components.default.messageContentView = DemoChatMessageContentView.self
+        Components.default.messageActionsVC = DemoChatMessageActionsVC.self
+        Components.default.reactionsSorting = { $0.type.position < $1.type.position }
+        Components.default.messageLayoutOptionsResolver = DemoChatMessageLayoutOptionsResolver()
+    }
+}
