+++ DemoApp/StreamChat/StreamChatWrapper.swift
@@ -14,7 +14,7 @@ final class StreamChatWrapper {
     var onRemotePushRegistration: (() -> Void)?
 
     // Chat client
-    private var client: ChatClient?
+    var client: ChatClient?
 
     // ChatClient config
     var config: ChatClientConfig = {
@@ -54,20 +54,6 @@ final class StreamChatWrapper {
                 ? Bundle.main.localizedString(forKey: key, value: nil, table: table)
                 : localizedString
         }
-    }
-
-    func configureUI() {
-        // Customize UI configuration
-        Components.default.messageListDateSeparatorEnabled = true
-        Components.default.messageListDateOverlayEnabled = true
-
-        // Customize UI components
-        Components.default.channelListRouter = DemoChatChannelListRouter.self
-        Components.default.channelVC = DemoChatChannelVC.self
-        Components.default.messageContentView = DemoChatMessageContentView.self
-        Components.default.messageActionsVC = DemoChatMessageActionsVC.self
-        Components.default.reactionsSorting = { $0.type.position < $1.type.position }
-        Components.default.messageLayoutOptionsResolver = DemoChatMessageLayoutOptionsResolver()
     }
 }
