--- StreamChatUITestsApp/StreamChat/ChatClient+shared.swift
@@ -1,10 +1,0 @@
-//
-// Copyright © 2021 Stream.io Inc. All rights reserved.
-//
-
-import Foundation
-import StreamChat
-
-extension ChatClient {
-    static var shared: ChatClient!
-}
